requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
channels==4.0.0
channels-redis==4.1.0
Django==5.0.6
djangorestframework==3.14.0
msgpack==1.0.8
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
sqlparse==0.5.0
